+++ features/step_definitions/cache.rb
@@ -1,4 +1,4 @@
-require 'xcodeproj'
+require "xcodeproj"
 
 Then(/^tuist warms the cache$/) do
   system("swift", "run", "tuist", "cache", "warm", "--path", @dir)