+++ features/support/assertions.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require 'minitest/spec'
-require 'minitest/assertions'
+require "minitest/spec"
+require "minitest/assertions"
 
 class MinitestWorld
   include Minitest::Assertions