+++ tools/fourier/test/test_helper.rb
@@ -4,23 +4,23 @@ addpath = lambda do |p|
 end
 addpath.call(File.expand_path("../lib", __dir__))
 
-require 'cli/ui'
-require 'fileutils'
-require 'tmpdir'
-require 'tempfile'
-require 'byebug'
+require "cli/ui"
+require "fileutils"
+require "tmpdir"
+require "tempfile"
+require "byebug"
 
 CLI::UI::StdoutRouter.enable
 
-require 'minitest/autorun'
+require "minitest/autorun"
 require "minitest/unit"
-require 'minitest/reporters'
+require "minitest/reporters"
 
 reporter_options = { color: true }
 Minitest::Reporters.use!([Minitest::Reporters::DefaultReporter.new(reporter_options)])
 
-require 'mocha/minitest'
-require 'fourier'
+require "mocha/minitest"
+require "fourier"
 
 class TestCase < MiniTest::Test
   def root_directory